+++ src/jstrimesh/trimeshPub.h
@@ -20,13 +20,13 @@
 
 struct Surface {
 
-	size_t vertexCount;
+	int vertexCount;
 	SURFACE_REAL_TYPE *vertex; // x,y,z,...
 	SURFACE_REAL_TYPE *normal; // x,y,z,... 
 	SURFACE_REAL_TYPE *textureCoordinate; // s,t,r,...
 	SURFACE_REAL_TYPE *color; // r,g,b,a,...
 
-	size_t indexCount;
+	int indexCount;
 	SURFACE_INDEX_TYPE *index;
 };
